requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
arxiv
wikipedia
langchain_groq
langchain_huggingface
langchain_chroma
ipykernel
langchain-text-splitters
ddgs
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
langchain_huggingface
huggingface_hub
gradio